code/requirements.txt
transformers
wandb
bitsandbytes
peft
datasets
torch
python-dotenv
pyyaml
accelerate